GitHub Actions (#1)
This automatically pushes images to the docker hub whenever:

* A new sharp release of Portainer has been made
* or the content of the main branch here changes

This means that there will be as many images in our target registry as there are original portainer images. The implementation relays the DockerHub-compatible hooks scripts found in the hooks sub-directory, making it possible to trigger the same kind of logic, but from the DockerHub infrastructure instead. The last step of the github actions is to update the description of the image at the DockerHub.